Web UI for creating domains

+++ priv/repo/migrations/20211027200445_create_domains.exs
@@ -5,7 +5,7 @@ defmodule Virt.Repo.Migrations.CreateDomains do
     create table(:domains, primary_key: false) do
       add :id, :binary_id, primary_key: true
       add :name, :string, null: false
-      add :memory_bytes, :integer, null: false
+      add :memory_bytes, :bigint, null: false
       add :vcpus, :integer, null: false
       add :created, :boolean, default: false
       add :host_id, references(:hosts, on_delete: :nothing, type: :binary_id)
